+++ packages/commercetools/api-client/src/api/createCart/index.ts
@@ -1,6 +1,6 @@
-import { CartDraft } from './../../types/GraphQL';
-import { apolloClient, getSettings } from './../../index';
-import { getCustomQuery } from './../../helpers/queries';
+import { CartDraft } from '../../types/GraphQL';
+import { CustomQueryFn, getSettings } from '../../index';
+import { getCustomQuery } from '../../helpers/queries';
 import defaultQuery from './defaultMutation';
 import gql from 'graphql-tag';
 
@@ -8,8 +8,8 @@ interface CartData extends Omit<CartDraft, 'currency'> {
   currency?: string;
 }
 
-const createCart = async (cartDraft: CartData = {}, customQueryFn?) => {
-  const { locale, acceptLanguage, currency } = getSettings();
+const createCart = async (cartDraft: CartData = {}, customQueryFn?: CustomQueryFn) => {
+  const { locale, acceptLanguage, currency, client } = getSettings();
 
   const defaultVariables = {
     acceptLanguage,
@@ -21,10 +21,8 @@ const createCart = async (cartDraft: CartData = {}, customQueryFn?) => {
   };
   const { query, variables } = getCustomQuery(customQueryFn, { defaultQuery, defaultVariables });
 
-  const request = await apolloClient.mutate({
-    mutation: gql`
-      ${query}
-    `,
+  const request = await client.mutate({
+    mutation: gql`${query}`,
     variables,
     fetchPolicy: 'no-cache'
   });