+++ packages/commercetools/theme/pages/Checkout/Shipping.vue
@@ -6,7 +6,7 @@
       class="sf-heading--left sf-heading--no-underline title"
     />
     <ValidationObserver v-slot="{ handleSubmit, dirty, reset }">
-      <form @submit.prevent="handleSubmit(dirty || addressIsModified ? handleShippingAddressSubmit(reset) : handleShippingMethodSubmit(reset))">
+      <form @submit.prevent="handleSubmit(canContinueToPayment(dirty) ? handleShippingMethodSubmit(reset) : handleShippingAddressSubmit(reset))">
         <UserShippingAddresses
           v-if="isAuthenticated && shippingAddresses && shippingAddresses.length"
           :setAsDefault="setAsDefault"
@@ -123,7 +123,7 @@
         </div>
         <SfButton
           v-if="!canAddNewAddress"
-          class="form__action-button form__action-button--margin-bottom"
+          class="color-light form__action-button form__action-button--add-address"
           type="submit"
           @click.native="canAddNewAddress = true"
         >
@@ -299,6 +299,7 @@ export default {
       addressIsModified.value = false;
     };
     const handleShippingMethodSubmit = (reset) => async () => {
+      await setShippingMethod(chosenShippingMethod.value, { save: true });
       reset();
       context.root.$router.push('/checkout/payment');
     };
@@ -306,7 +307,7 @@ export default {
     const setShippingDetailsAndUnpickAddress = value => {
       setShippingDetails(value);
       currentAddressId.value = -1;
-      addressIsModified.value = false;
+      addressIsModified.value = true;
     };
 
     const canContinueToPayment = dirty => isShippingAddressCompleted.value && !dirty && !addressIsModified.value;
@@ -342,15 +343,26 @@ export default {
 
 .title {
   margin: var(--spacer-xl) 0 var(--spacer-base) 0;
-  @include for-desktop {
-    margin: var(--spacer-2xl) 0 var(--spacer-base) 0;
-  }
 }
 .form {
+  --button-width: 100%;
+  &__select {
+    display: flex;
+    align-items: center;
+    --select-option-font-size: var(--font-size--lg);
+    ::v-deep .sf-select__dropdown {
+      font-size: var(--font-size--lg);
+      margin: 0;
+      color: var(--c-text);
+      font-family: var(--font-family--secondary);
+      font-weight: var(--font-weight--normal);
+    }
+  }
   @include for-desktop {
     display: flex;
     flex-wrap: wrap;
     align-items: center;
+    --button-width: auto;
   }
   &__element {
     margin: 0 0 var(--spacer-xl) 0;
@@ -382,30 +394,37 @@ export default {
     &--secondary {
       @include for-desktop {
         order: -1;
-        --button-margin: 0;
         text-align: left;
+      }
+    }
+    &--add-address {
+      width: 100%;
+      margin: 0;
+      @include for-desktop {
+        margin: 0 0 var(--spacer-lg) 0;
+        width: auto;
       }
     }
   }
   &__back-button {
-    margin: 0 var(--spacer-xl) 0 0;
+    margin: var(--spacer-xl) 0 var(--spacer-sm);
     &:hover {
-      color:  white;
+      color:  var(--c-white);
     }
-  }
-  &__button {
-    --button-width: 100%;
     @include for-desktop {
-      --button-width: auto;
+      margin: 0 var(--spacer-xl) 0 0;
     }
   }
   &__radio-group {
     flex: 0 0 100%;
-    margin: 0 0 var(--spacer-2xl) 0;
+    margin: 0 0 var(--spacer-xl) 0;
+    @include for-desktop {
+      margin: 0 0 var(--spacer-2xl) 0;
+    }
+
   }
 }
 .shipping {
-  margin: 0 calc(var(--spacer-xl) * -1);
   &__label {
     display: flex;
     justify-content: space-between;
@@ -413,22 +432,6 @@ export default {
   &__description {
     --radio-description-margin: 0;
     --radio-description-font-size: var(--font-xs);
-  }
-  &__delivery {
-    color: var(--c-text-muted);
-  }
-  &__action {
-    margin: 0 0 0 var(--spacer);
-    &::before {
-      content: "+";
-    }
-    &--is-active {
-      --button-color: var(--c-primary);
-      --button-transition: color 150ms linear;
-      &::before {
-        content: "-";
-      }
-    }
   }
 }
 </style>