+++ packages/core/nuxt-theme-module/theme/components/BottomNavigation.vue
@@ -1,17 +1,18 @@
 <template>
 <!-- TODO: create logic with isActive prop for BottomNavigationItems -->
-  <SfBottomNavigation class="mobile-only">
+  <SfBottomNavigation class="smartphone-only">
     <nuxt-link data-cy="bottom-navigation-url_home" to="/">
       <SfBottomNavigationItem :class="$route.path == '/' ? 'sf-bottom-navigation__item--active' : ''" icon="home" size="20px" label="Home"/>
     </nuxt-link>
     <SfBottomNavigationItem data-cy="bottom-navigation-url_menu" icon="menu" size="20px" label="Menu"/>
-    <SfBottomNavigationItem data-cy="bottom-navigation-url_wishlist" icon="heart" size="20px" label="Wishlist"/>
-    <SfBottomNavigationItem data-cy="bottom-navigation-url_account" icon="profile" size="20px" label="Account"/>
+    <SfBottomNavigationItem data-cy="bottom-navigation-url_wishlist" icon="heart" size="20px" label="Wishlist" @click="toggleWishlistSidebar"/>
+    <SfBottomNavigationItem data-cy="bottom-navigation-url_account" icon="profile" size="20px" label="Account" @click="handleAccountClick"/>
     <!-- TODO: add logic for label - if on Home then Basket, if on PDC then AddToCart etc. -->
     <SfBottomNavigationItem data-cy="bottom-navigation-url_add-to-cart"
       label="Basket"
       icon="add_to_cart"
-      >
+      @click="toggleCartSidebar"
+    >
       <template #icon>
         <SfCircleIcon aria-label="Add to cart">
           <SfIcon
@@ -28,12 +29,31 @@
 
 <script>
 import { SfBottomNavigation, SfIcon, SfCircleIcon } from '@storefront-ui/vue';
+import { useUiState } from '~/composables';
+import { useUser } from '<%= options.generate.replace.composables %>';
 
 export default {
   components: {
     SfBottomNavigation,
     SfIcon,
     SfCircleIcon
+  },
+  setup(props, { root }) {
+    const { toggleCartSidebar, toggleWishlistSidebar, toggleLoginModal } = useUiState();
+    const { isAuthenticated } = useUser();
+
+    const handleAccountClick = async () => {
+      if (isAuthenticated.value) {
+        return root.$router.push('/my-account');
+      }
+      toggleLoginModal();
+    };
+
+    return {
+      toggleWishlistSidebar,
+      toggleCartSidebar,
+      handleAccountClick
+    };
   }
 };
 </script>