+++ packages/core/docs/commercetools/changelog.md
@@ -1,9 +1,15 @@
 # Changelog
 
-## 0.2.5 - not released
+## 0.2.6 - not released
 
+## 0.2.5
+
+- `customQuery` for checkout composables ([#5025](https://github.com/DivanteLtd/vue-storefront/issues/5025))
+- api-client apollo client no longer shared between requests ([#5056](https://github.com/DivanteLtd/vue-storefront/pull/5056))
 
 ## 0.2.4
+
+- Remove defaulting for checkout shipping details ([#5026](https://github.com/DivanteLtd/vue-storefront/issues/5026))
 
 ### Changes